Reorganize code according to comments
